ai/requirement.txt
fastapi==0.115.11
uvicorn
aiofiles
pdfplumber
python-docx
httpx
